image slider updated
