+++ src/components/project/index.ts
@@ -1,3 +1,4 @@
 export { ProjectCard } from "./card/project-card";
 export { ProjectBanner } from "./banner/project-banner";
 export { ProjectDescription } from "./description/project-description";
+export { ProjectImageSlider } from "./project-image-slider/project-image-slider";
