+++ src/components/project/project-image-slider/project-image-slider.module.css
@@ -1,0 +1,150 @@
+.root {
+  position: relative;
+  margin-left: auto;
+  margin-right: auto;
+  width: 83.333%;
+  padding-top: 5rem;
+  padding-bottom: 5rem;
+
+  @media (min-width: 1024px) {
+    padding-top: 10rem;
+    padding-bottom: 10rem;
+  }
+
+  @media (min-width: 1280px) {
+    padding-top: 0;
+    padding-bottom: 0;
+  }
+
+  @media (min-width: 1792px) {
+    margin-left: 0;
+    margin-right: 0;
+  }
+}
+
+.section {
+  display: flex;
+  width: 100%;
+  max-width: 100%;
+  flex-direction: column;
+  gap: 1.5rem;
+  padding-left: 1.5rem;
+  padding-right: 1.5rem;
+
+  @media (min-width: 640px) {
+    padding-left: 0;
+    padding-right: 0;
+  }
+}
+
+.title {
+  font-family: var(--font-heading);
+  font-size: 1.5rem;
+  font-weight: 600;
+  line-height: 1;
+
+  @media (min-width: 640px) {
+    font-size: 1.875rem;
+  }
+
+  @media (min-width: 768px) {
+    font-size: 2.25rem;
+  }
+
+  @media (min-width: 1280px) {
+    font-size: 3rem;
+  }
+}
+
+.image-container {
+  position: relative;
+  display: flex;
+  aspect-ratio: 16 / 9;
+  cursor: pointer;
+  align-items: center;
+  justify-content: center;
+  overflow: hidden;
+  border-radius: 0.375rem;
+
+  &:hover .icon-hover {
+    transform: scale(1);
+    opacity: 1;
+  }
+
+  @media (min-width: 1280px) {
+    &.background {
+      aspect-ratio: 4 / 3;
+    }
+  }
+}
+
+.icon-hover {
+  position: absolute;
+  right: 0.5rem;
+  bottom: 0.5rem;
+  display: none;
+  cursor: pointer;
+  align-items: center;
+  justify-content: center;
+  border-radius: 9999px;
+  background-color: white;
+  padding: 0.25rem;
+  color: black;
+  box-shadow: 0 4px 6px -1px rgb(0 0 0 / 0.1);
+  transition-property: all;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 300ms;
+  transform: scale(0);
+  opacity: 0;
+
+  @media (min-width: 640px) {
+    right: 0.75rem;
+    bottom: 0.75rem;
+    padding: 0.5rem;
+  }
+
+  @media (min-width: 1024px) {
+    display: flex;
+  }
+}
+
+.icon {
+  height: 18px;
+  width: 18px;
+}
+
+.modal {
+  position: fixed;
+  inset: 0;
+  z-index: 50;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: rgb(0 0 0 / 0.7);
+  transition-property: opacity;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 300ms;
+}
+
+.modal-content {
+  position: relative;
+  width: 83.333%;
+  aspect-ratio: 16 / 9;
+
+  @media (min-width: 640px) {
+    width: 91.666%;
+  }
+
+  @media (min-width: 1280px) {
+    width: 75%;
+  }
+
+  @media (min-width: 1792px) {
+    width: 66.666%;
+  }
+}
+
+.modal-image {
+  cursor: pointer;
+  object-fit: contain;
+}
