+++ src/components/ui/accordion/accordion.module.css
@@ -21,6 +21,7 @@
   font-family: var(--font-heading);
   font-size: 1.5rem;
   font-weight: 600;
+  line-height: 1;
 
   @media (min-width: 640px) {
     font-size: 1.875rem;
